web_operations.py
import requests
from bs4 import BeautifulSoup
import json
from typing import Dict, List, Any, Optional
import os
import youtube_dl
import urllib.parse
from datetime import datetime
import aiohttp
import asyncio
from concurrent.futures import ThreadPoolExecutor
import re

class WebOperations:
    """
    A class to handle various web operations including searching,
    downloading content, and web scraping.
    """
    
    def search_web(self, query: str) -> str:
        """
        Perform a web search.
        
        Args:
            query (str): Search query
            
        Returns:
            str: Search results or error message
        """
        try:
            # This is a placeholder. In a real implementation, you would use a search API
            return f"Search results for: {query}"
        except Exception as e:
            return f"Error performing web search: {str(e)}"

    def search_images(self, query: str) -> str:
        """
        Search for images.
        
        Args:
            query (str): Search query
            
        Returns:
            str: Image search results or error message
        """
        try:
            # This is a placeholder. In a real implementation, you would use an image search API
            return f"Image search results for: {query}"
        except Exception as e:
            return f"Error searching images: {str(e)}"

    def search_videos(self, query: str) -> str:
        """
        Search for videos.
        
        Args:
            query (str): Search query
            
        Returns:
            str: Video search results or error message
        """
        try:
            # This is a placeholder. In a real implementation, you would use a video search API
            return f"Video search results for: {query}"
        except Exception as e:
            return f"Error searching videos: {str(e)}"

    def search_news(self, query: str) -> str:
        """
        Search for news articles.
        
        Args:
            query (str): Search query
            
        Returns:
            str: News search results or error message
        """
        try:
            # This is a placeholder. In a real implementation, you would use a news API
            return f"News search results for: {query}"
        except Exception as e:
            return f"Error searching news: {str(e)}"

    def search_academic(self, query: str) -> str:
        """
        Search for academic papers.
        
        Args:
            query (str): Search query
            
        Returns:
            str: Academic search results or error message
        """
        try:
            # This is a placeholder. In a real implementation, you would use an academic search API
            return f"Academic search results for: {query}"
        except Exception as e:
            return f"Error searching academic papers: {str(e)}"

    def download_file(self, url: str, path: str) -> str:
        """
        Download a file from a URL.
        
        Args:
            url (str): URL of the file to download
            path (str): Local path to save the file
            
        Returns:
            str: Success or error message
        """
        try:
            response = requests.get(url, stream=True)
            response.raise_for_status()
            
            with open(path, 'wb') as f:
                for chunk in response.iter_content(chunk_size=8192):
                    f.write(chunk)
            
            return f"File downloaded successfully to {path}"
        except Exception as e:
            return f"Error downloading file: {str(e)}"

    def download_image(self, url: str, path: str) -> str:
        """
        Download an image from a URL.
        
        Args:
            url (str): URL of the image to download
            path (str): Local path to save the image
            
        Returns:
            str: Success or error message
        """
        try:
            response = requests.get(url, stream=True)
            response.raise_for_status()
            
            # Verify that the content is an image
            content_type = response.headers.get('content-type', '')
            if not content_type.startswith('image/'):
                return f"Error: URL does not point to an image (content-type: {content_type})"
            
            with open(path, 'wb') as f:
                for chunk in response.iter_content(chunk_size=8192):
                    f.write(chunk)
            
            return f"Image downloaded successfully to {path}"
        except Exception as e:
            return f"Error downloading image: {str(e)}"

    def download_video(self, url: str, path: str) -> str:
        """
        Download a video from a URL.
        
        Args:
            url (str): URL of the video to download
            path (str): Local path to save the video
            
        Returns:
            str: Success or error message
        """
        try:
            response = requests.get(url, stream=True)
            response.raise_for_status()
            
            # Verify that the content is a video
            content_type = response.headers.get('content-type', '')
            if not content_type.startswith('video/'):
                return f"Error: URL does not point to a video (content-type: {content_type})"
            
            with open(path, 'wb') as f:
                for chunk in response.iter_content(chunk_size=8192):
                    f.write(chunk)
            
            return f"Video downloaded successfully to {path}"
        except Exception as e:
            return f"Error downloading video: {str(e)}"

# Create a singleton instance
web_operations = WebOperations() 