requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver-manager==4.0.1
numpy==1.24.4
opencv-python==4.8.1.78
moviepy==1.0.3
soundfile==0.12.1
librosa==0.10.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2
pandas==2.0.3
sqlalchemy==1.4.50
mysql-connector-python==8.3.0
psycopg==3.1.18
pymongo==4.6.1
pyyaml==6.0.1
openpyxl==3.1.2
websocket-client==1.7.0
sounddevice==0.4.6
psutil==5.9.8
schedule==1.2.1
python-docx==1.1.0
markdown==3.5.2
pdfkit==1.0.0
wkhtmltopdf==0.2
youtube-dl==2021.12.17
aiohttp==3.9.3
lxml==5.1.0
paramiko==3.4.0
docker==7.0.0
gitpython==3.1.42
watchdog==3.0.0
keyboard==0.13.5; platform_system != "Darwin"
mouse==0.7.1; platform_system != "Darwin"
pyautogui==0.9.53
python-xlib==0.33; platform_system == "Linux"
pywin32==306; platform_system == "Windows"
websocket-client==1.7.0
--find-links https://github.com/python-pillow/Pillow/releases/download/9.5.0/
Pillow==9.5.0; sys_platform == 'darwin'
Pillow==10.1.0; sys_platform != 'darwin'
websockets==12.0
